pyproject.toml
[project]
name = "researchbuddy"
version = "0.1.4"
description = "AI-powered review research assistant with parallel crawling and synthesis"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pydantic>=2.9.0",
    "pydantic-settings>=2.6.0",
    "aiosqlite>=0.20.0",
    "httpx>=0.28.0",
    "transformers>=4.51.0",
    "openai-whisper>=20250625",
    "praw>=7.7.1",
    "pypdf>=5.4.0",
    "python-dotenv>=1.0.1",
    "tiktoken>=0.12.0",
    "typer>=0.9.0",
    "rich>=13.0.0",
    "playwright>=1.50.0",
    "crawl4ai>=0.7.0",
    "yt-dlp>=2025.12.8",
    "tokonomics>=1.2.14 ; python_full_version >= '3.13'",
]

[project.scripts]
researchbuddy = "app.cli:main"

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-mock>=3.14.0",
    "ruff>=0.8.0",
    "httpx>=0.28.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "C4", "SIM"]
ignore = ["E501"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
testpaths = ["app/tests"]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytest-mock>=3.15.1",
    "ruff>=0.14.8",
]
